Extract a number of helper functions into a helpers.py file
